requirements.txt
pip == 22.0.2
discord.py == 2.1.0
python_dotenv==0.21.1